Extracted readers and writers to io related modules
